feat:add error handling in all routes

+++ server/controllers/blogData.js
@@ -11,7 +11,10 @@ const getAllData=asyncWrapper(async(req,res)=>{
 
 
 
+
 const createBlog=asyncWrapper(async(req,res)=>{
+    const {blogTitle,blogData,authorId,authorName,imageUrl}=req.body;
+    if(!blogTitle || !blogData || !authorId || !authorName || !imageUrl) throw new CustomError("Bad Request",StatusCodes.BAD_REQUEST);
     const response=await Blog.create(req.body);
     res.status(StatusCodes.OK).json(response);
 })
@@ -19,6 +22,7 @@ const createBlog=asyncWrapper(async(req,res)=>{
 
 const getBlogById=asyncWrapper(async(req,res)=>{
     const {id}=req.params;
+    if(!id) throw new CustomError("Bad Request",StatusCodes.BAD_REQUEST);
     const response=await Blog.findById(id);
     if(!response) throw new CustomError("Data not found",StatusCodes.NOT_FOUND);
     res.status(StatusCodes.OK).json(response)
@@ -27,6 +31,7 @@ const getBlogById=asyncWrapper(async(req,res)=>{
 
 const getBlogByAuthorId=asyncWrapper(async(req,res)=>{
     const {userId:authorId}=req.body;
+    if(!authorId) throw new CustomError("Bad Request",StatusCodes.BAD_REQUEST);
     const response=await Blog.find({authorId});
     if(!response){
         res.status(StatusCodes.OK).json({
@@ -40,13 +45,14 @@ const getBlogByAuthorId=asyncWrapper(async(req,res)=>{
 
 const editBlog=asyncWrapper(async(req,res)=>{
     const {id:_id}=req.params;
+    if(!_id) throw new CustomError("Bad Request",StatusCodes.BAD_REQUEST);
     const updatedData=await Blog.findOneAndUpdate({_id},req.body,{runValidators:true,new:true});
-    console.log(updatedData);
     res.status(StatusCodes.OK).json(updatedData);
 })
 
 const deleteBlog=asyncWrapper(async(req,res)=>{
     const {id:_id}=req.params;
+    if(!_id) throw new CustomError("Bad Request",StatusCodes.BAD_REQUEST);
     const response=await Blog.findOneAndDelete({_id});
     res.status(StatusCodes.OK).json(response);
